scripts/train/requirements.txt
# Flyeas V7 Training Pipeline — Python dependencies
# Install: pip install -r scripts/train/requirements.txt

# Env loading (CRITICAL — loads .env.local)
python-dotenv>=1.0.0

# Core ML (CPU-only, M1/M2 compatible)
torch>=2.1.0
onnx>=1.15.0
onnxruntime>=1.16.0

# Data processing
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.25.0

# ML algorithms
scikit-learn>=1.3.0
scipy>=1.11.0

# Reports
matplotlib>=3.8.0
tabulate>=0.9.0

# DB
supabase>=2.0.0
